+++ base-hack/asm/objects.asm
@@ -21,6 +21,15 @@
 .importobj "obj/src_initialization_stack_trace.o"
 .importobj "obj/src_initialization_text.o"
 .importobj "obj/src_initialization_widescreen.o"
+.importobj "obj/src_instances_bananaport.o"
+.importobj "obj/src_instances_boss_portal.o"
+.importobj "obj/src_instances_cannon.o"
+.importobj "obj/src_instances_crown_pad.o"
+.importobj "obj/src_instances_instances.o"
+.importobj "obj/src_instances_instance_lib.o"
+.importobj "obj/src_instances_melon_crate.o"
+.importobj "obj/src_instances_shops.o"
+.importobj "obj/src_instances_switches.o"
 .importobj "obj/src_item rando_actors.o"
 .importobj "obj/src_item rando_archipelago.o"
 .importobj "obj/src_item rando_collision.o"
@@ -43,7 +52,6 @@
 .importobj "obj/src_misc_file_screen.o"
 .importobj "obj/src_misc_hard_mode.o"
 .importobj "obj/src_misc_helm_hurry.o"
-.importobj "obj/src_misc_instances.o"
 .importobj "obj/src_misc_item_db.o"
 .importobj "obj/src_misc_krusha.o"
 .importobj "obj/src_misc_move_hints.o"
