+++ base-hack/assets/instance_scripts/isles/trombone.json
@@ -1,0 +1,85 @@
+{
+    "id": 49,
+    "behav_9C": 0,
+    "ignore": false,
+    "output_version": 2,
+    "script": [
+        {
+            "conditions": [
+                {"function":1,"inverted":false,"parameters":[0,0,0]},
+                {"function":45,"inverted":false,"parameters":[426,0,0]}
+            ],
+            "executions": [
+                {"function":38,"parameters":[2,0,0]}
+            ]
+        },
+        {
+            "conditions": [
+                {"function":1,"inverted":false,"parameters":[0,0,0]},
+                {"function":45,"inverted":true,"parameters":[426,0,0]}
+            ],
+            "executions": [
+                {"function":1,"parameters":[1,0,0]}
+            ]
+        },
+        {
+            "conditions": [
+                {"function":1,"inverted":false,"parameters":[1,0,0]},
+                {"function":17,"inverted":true,"parameters":[61,0,0]}
+            ],
+            "executions": [
+                {"function":1,"parameters":[2,0,0]}
+            ]
+        },
+        {
+            "conditions": [
+                {"function":1,"inverted":false,"parameters":[2,0,0]},
+                {"function":17,"inverted":false,"parameters":[61,0,0]}
+            ],
+            "executions": [
+                {"function":1,"parameters":[1,0,0]}
+            ]
+        },
+        {
+            "conditions": [
+                {"function":1,"inverted":false,"parameters":[2,0,0]},
+                {"function":13,"inverted":false,"parameters":[2,0,0]},
+                {"function":6,"inverted":false,"parameters":[7,65521,0]},
+                {"function":23,"inverted":false,"parameters":[103,0,0]}
+            ],
+            "executions": [
+                {"function":38,"parameters":[1,0,0]},
+                {"function":1,"parameters":[3,0,0]}
+            ]
+        },
+        {
+            "conditions": [
+                {"function":1,"inverted":false,"parameters":[3,0,0]},
+                {"function":35,"inverted":true,"parameters":[0,0,0]}
+            ],
+            "executions": [
+                {"function":37,"parameters":[14,1,0]},
+                {"function":1,"parameters":[4,0,0]}
+            ]
+        },
+        {
+            "conditions": [
+                {"function":1,"inverted":false,"parameters":[5,0,0]}
+            ],
+            "executions": [
+                {"function":94,"parameters":[9,0,0]},
+                {"function":1,"parameters":[6,0,0]}
+            ]
+        },
+        {
+            "conditions": [
+                {"function":1,"inverted":false,"parameters":[6,0,0]},
+                {"function":35,"inverted":true,"parameters":[0,0,0]}
+            ],
+            "executions": [
+                {"function":107,"parameters":[426,1,0]},
+                {"function":38,"parameters":[2,0,0]}
+            ]
+        }
+    ]
+}
