+++ base-hack/include/global.h
@@ -416,6 +416,22 @@ extern void handleGrabbingLock(void* player, int player_index, int allow_vines);
 extern void handleActionSet(int action, void* actor, int player_index);
 extern int getTrackerYOffset(void);
 
+extern void hideObject(behaviour_data* behaviour_pointer);
+extern void shopGenericCode(behaviour_data* behaviour, int index, int id, vendors shop);
+extern void bananaportGenericCode(behaviour_data* behaviour, int index, int id);
+extern void TNSPortalGenericCode(behaviour_data* behaviour, int index, int id);
+extern void TNSIndicatorGenericCode(behaviour_data* behaviour, int index, int id);
+extern void CrownPadGenericCode(behaviour_data* behaviour, int index, int id, int crown_level_index);
+extern void MelonCrateGenericCode(behaviour_data* behaviour, int index, int id);
+extern int isBonus(maps map);
+extern int randomGunSwitchGenericCode(behaviour_data* behaviour_pointer, int index, int switch_index);
+extern int randomInstrumentGenericCode(behaviour_data* behaviour_pointer, int index, int pad_index);
+extern int checkControlState(int target_control_state);
+extern int checkSlamLocation(int kong, int key, int id);
+extern void playSFXContainer(int id, int vanilla_sfx, int new_sfx);
+extern int getPressedSwitch(behaviour_data* behaviour_pointer, int bullet_type, int ID);
+extern void getModelTwoItemFromActor(int actor, short* item, float* scale);
+
 extern unsigned int cs_skip_db[432];
 extern bonus_barrel_info bonus_data[95];
 extern const short kong_flags[5];
@@ -423,6 +439,7 @@ extern const short normal_key_flags[8];
 extern const unsigned short slam_flags[6];
 extern const unsigned short belt_flags[4];
 extern const unsigned short instrument_flags[6];
+extern const unsigned char kong_pellets[5];
 extern const rgb colorblind_colors[15];
 extern const check_struct item_db[292];
 extern const unsigned char crown_maps[10];
